Add possobility to update liquibase and postgreDB through the start.sh script, create 2 profiles for it and create a simple implementation of MongoRepository

+++ menu-common/web/src/main/resources/db/changelog/postgre/changes/changeset_1.0.0.sql
@@ -18,7 +18,7 @@ CREATE TABLE menu (
     id BIGINT NOT NULL PRIMARY KEY DEFAULT nextval('global_seq'),
     rate FLOAT,
     creating_date TIMESTAMP NOT NULL,
-    restaurant_id INT NOT NULL,
+    restaurant_id BIGINT NOT NULL,
     FOREIGN KEY (restaurant_id) REFERENCES restaurant(id)
 );
